move ResultChecker into .helper
